requirements.txt
google-api-python-client
newspaper3k
lxml_html_clean
ollama
langchain
langchain-community
sentence_transformers
faiss-cpu